+++ dizoo/d4rl/config/hopper_medium_expert_pd_config.py
@@ -1,9 +1,9 @@
 from easydict import EasyDict
 
 main_config = dict(
-    exp_name="halfcheetah_medium_pd_seed0",
+    exp_name="hopper_medium_expert_pd_seed0",
     env=dict(
-        env_id='halfcheetah-medium-v2',
+        env_id='hopper-medium-expert-v2',
         collector_env_num=1,
         evaluator_env_num=8,
         use_act_scale=True,
@@ -15,9 +15,9 @@ main_config = dict(
         include_returns=True,
         normed=False,
         stop_value=8000,
-        horizon=4,
-        obs_dim=17,
-        action_dim=6,
+        horizon=32,
+        obs_dim=11,
+        action_dim=3,
     ),
     policy=dict(
         cuda=True,
@@ -26,16 +26,16 @@ main_config = dict(
             diffuser_model_cfg=dict(
                 model='TemporalUnet',
                 model_cfg=dict(
-                    transition_dim=23,
+                    transition_dim=14,
                     dim=32,
-                    dim_mults=[1, 4, 8],
+                    dim_mults=[1, 2, 4, 8],
                     returns_condition=False,
                     kernel_size=5,
-                    attention=True,
+                    attention=False,
                 ),
-                horizon=4,
-                obs_dim=17,
-                action_dim=6,
+                horizon=32,
+                obs_dim=11,
+                action_dim=3,
                 n_timesteps=20,
                 predict_epsilon=False,
                 loss_discount=1,
@@ -45,21 +45,21 @@ main_config = dict(
             value_model_cfg=dict(
                 model='TemporalValue',
                 model_cfg=dict(
-                    horizon = 4,
-                    transition_dim=23,
+                    horizon = 32,
+                    transition_dim=14,
                     dim=32,
-                    dim_mults=[1, 4, 8],
+                    dim_mults=[1, 2, 4, 8],
                     kernel_size=5,
                 ),
-                horizon=4,
-                obs_dim=17,
-                action_dim=6,
+                horizon=32,
+                obs_dim=11,
+                action_dim=3,
                 n_timesteps=20,
                 predict_epsilon=True,
                 loss_discount=1,
             ),
             n_guide_steps=2,
-            scale=0.001,
+            scale=0.0001,
             t_stopgrad=4,
             scale_grad_by_std=True,
         ),
