+++ dizoo/d4rl/config/walker2d_medium_pd_config.py
@@ -1,9 +1,9 @@
 from easydict import EasyDict
 
 main_config = dict(
-    exp_name="halfcheetah_medium_pd_seed0",
+    exp_name="walker2d_medium_pd_seed0",
     env=dict(
-        env_id='halfcheetah-medium-v2',
+        env_id='walker2d-medium-v2',
         collector_env_num=1,
         evaluator_env_num=8,
         use_act_scale=True,
@@ -15,7 +15,7 @@ main_config = dict(
         include_returns=True,
         normed=False,
         stop_value=8000,
-        horizon=4,
+        horizon=32,
         obs_dim=17,
         action_dim=6,
     ),
@@ -28,12 +28,12 @@ main_config = dict(
                 model_cfg=dict(
                     transition_dim=23,
                     dim=32,
-                    dim_mults=[1, 4, 8],
+                    dim_mults=[1, 2, 4, 8],
                     returns_condition=False,
                     kernel_size=5,
-                    attention=True,
+                    attention=False,
                 ),
-                horizon=4,
+                horizon=32,
                 obs_dim=17,
                 action_dim=6,
                 n_timesteps=20,
@@ -45,13 +45,13 @@ main_config = dict(
             value_model_cfg=dict(
                 model='TemporalValue',
                 model_cfg=dict(
-                    horizon = 4,
+                    horizon = 32,
                     transition_dim=23,
                     dim=32,
-                    dim_mults=[1, 4, 8],
+                    dim_mults=[1, 2, 4, 8],
                     kernel_size=5,
                 ),
-                horizon=4,
+                horizon=32,
                 obs_dim=17,
                 action_dim=6,
                 n_timesteps=20,
@@ -59,8 +59,8 @@ main_config = dict(
                 loss_discount=1,
             ),
             n_guide_steps=2,
-            scale=0.001,
-            t_stopgrad=4,
+            scale=0.1,
+            t_stopgrad=2,
             scale_grad_by_std=True,
         ),
         normalizer='GaussianNormalizer',
